pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-cli"
version = "1.0.0"
description = "LLM CLI Chat — терминальный клиент для OpenRouter API с бенчмарком моделей"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "rich>=13.0",
    "httpx>=0.27",
    "pydantic>=2.0",
    "platformdirs>=4.0",
    "mcp[cli]>=1.0",
    "sentence-transformers>=3.0",
    "faiss-cpu>=1.8",
    "numpy>=1.26",
]

[project.scripts]
llm-cli = "llm_cli.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]
